+++ velox/connectors/clp/search_lib/BaseClpCursor.h
@@ -60,6 +60,43 @@ struct Field {
   std::string name;
 };
 
+enum class TimestampPrecision : uint8_t {
+  Seconds,
+  Milliseconds,
+  Microseconds,
+  Nanoseconds
+};
+
+/// Estimates the precision of an epoch timestamp as seconds, milliseconds,
+/// microseconds, or nanoseconds.
+///
+/// This heuristic relies on the fact that 1 year of epoch nanoseconds is
+/// approximately 1000 years of epoch microseconds and so on. This heuristic
+/// can be unreliable for timestamps sufficiently close to the epoch, but
+/// should otherwise be accurate for the next 1000 years.
+///
+/// Note: Future versions of the clp-s archive format will adopt a
+/// nanosecond-precision integer timestamp format (as opposed to the current
+/// format which allows other precisions), at which point we can remove this
+/// heuristic.
+///
+/// @param timestamp
+/// @return the estimated timestamp precision
+template <typename T>
+auto estimatePrecision(T timestamp) -> TimestampPrecision;
+
+/// Converts a float value into a Velox timestamp.
+///
+/// @param timestamp the input timestamp as a float
+/// @return the corresponding Velox timestamp
+auto convertToVeloxTimestamp(double timestamp) -> Timestamp;
+
+/// Converts an integer value into a Velox timestamp.
+///
+/// @param timestamp the input timestamp as an integer
+/// @return the corresponding Velox timestamp
+auto convertToVeloxTimestamp(int64_t timestamp) -> Timestamp;
+
 /// A query execution interface that manages the lifecycle of a query on a CLP-S
 /// split (archive or IR), including parsing and validating the query, loading
 /// the relevant splits, applying filters, and iterating over the results. It
